literate: show inline hints
